Move xservice into src, update script
