+++ site/package.json
@@ -18,7 +18,7 @@
     "storybook:build": "build-storybook",
     "test": "jest --selectProjects test",
     "test:coverage": "jest --selectProjects test --collectCoverage",
-    "typegen": "xstate typegen '**/*.ts'"
+    "typegen": "xstate typegen 'src/**/*.ts'"
   },
   "dependencies": {
     "@material-ui/core": "4.9.4",
